- Fix some warning on GCC (By Robert)
git-svn-id: https://svn.code.sf.net/p/speed-dreams/code/trunk@6823 30fe4595-0a0c-4342-8851-515496e4dcbd

Former-commit-id: 8c9f14fd2904cf6dd4abc423fc49436421e9cc74
Former-commit-id: a57281eb8159c44840daf39a7e76377e66824d1c

+++ src/drivers/shadow/src/Stuck.cpp
@@ -415,7 +415,7 @@ void	Stuck::init( const MyTrack& track, const tSituation* s, const tCarElt* me )
     // make opponent car cells unavailable.
     makeOpponentsList( s, me, &_opponents );
 
-    for( int i = 0; i < _opponents.size(); i++ )
+    for( size_t i = 0; i < _opponents.size(); i++ )
     {
         const OppInfo& oppInfo = _opponents[i];
         const tCarElt* other   = oppInfo.car;
@@ -1099,7 +1099,7 @@ void Stuck::dumpGrid() const
 {
     set<unsigned int> pts;
 
-    for( int i = 0; i < _plan.size(); i++ )
+    for( size_t i = 0; i < _plan.size(); i++ )
     {
         const GridPoint& pt = _plan[i];
         pts.insert( pt.x() * 256 + pt.y() );
